sdds/TA/requirements.txt
pytest-html
robotframework
artifactory
psutil
dirsync
distro
sseclient
websockets >= 11.0.3